+++ src/main/java/kitchenpos/order/domain/OrderLineItems.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.order.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
