+++ src/main/java/kitchenpos/order/domain/OrderRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.order.domain;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
