+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.order.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -14,22 +14,22 @@ import java.util.Optional;
 import kitchenpos.menu.fixture.MenuFixture;
 import kitchenpos.menu.fixture.MenuGroupFixture;
 import kitchenpos.menu.fixture.MenuProductFixture;
-import kitchenpos.application.fixture.OrderFixture;
-import kitchenpos.application.fixture.OrderLineItemFixture;
-import kitchenpos.application.fixture.ProductFixture;
+import kitchenpos.order.fixture.OrderFixture;
+import kitchenpos.order.fixture.OrderLineItemFixture;
+import kitchenpos.product.fixture.ProductFixture;
 import kitchenpos.table.fixture.TableFixture;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderLineItemRepository;
-import kitchenpos.domain.OrderRepository;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItemRepository;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 import kitchenpos.domain.Quantity;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
