+++ src/test/java/kitchenpos/order/fixture/OrderFixture.java
@@ -1,9 +1,9 @@
-package kitchenpos.application.fixture;
+package kitchenpos.order.fixture;
 
 import java.util.List;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 
 public class OrderFixture {
